server/requirements.txt
websockets==10.1
aioconsole==0.3.3
torchvision==0.11.1